Fix type of shared.Food().

+++ snakeDemo/src/shared/Food.java
@@ -8,7 +8,7 @@ public class Food extends PVector {
     static final Random random = new Random();
     private static final long serialVersionUID = 43L;
 
-    public Food(int x, int y) { super(x, y); }
+    public Food(float x, float y) { super(x, y); }
 
     public static Food randomFood(int fieldWidth, int fieldHeight) {
         return new Food(
